deletes gallery from subdomain

+++ controllers/galleries.js
@@ -2,7 +2,8 @@ const Gallery = require('../models/gallery');
 const response = require('../services/response');
 const S3 = require('../services/s3');
 const Image = require('../models/image');
-const { addGallery } = require('./subdomains');
+const { addGallery, deleteGallery } = require('./subdomains');
+const Subdomain = require('../models/subdomain');
 
 
 
@@ -53,7 +54,6 @@ const galleriesController = {
         { $pull: { images: imageId } },
         { new: true }
       );
-      console.log("gallery c", gallery)
       return gallery;
     } catch (e) {
       res.json(response.buildError(e));
@@ -76,12 +76,20 @@ const galleriesController = {
   delete: async (req, res) => {
     try {
       const { id } = req.params;
+      const subdomain = await Subdomain.findOne({ galleries: id })
       const gallery = await Gallery.findById({ _id: id });
-      const images = await Image.find({ "_id": { $in: gallery.images }})
-      S3.delete(images.map(img => img.image_url))
-      await Image.deleteMany({ "_id": { $in: gallery.images }})
-      await Gallery.deleteOne({ _id: gallery._id });
-      res.json(response.deleteGallery(`Successfuly deleted ${gallery.name} ` ))
+      if(gallery) {
+        const images = await Image.find({ "_id": { $in: gallery.images }})
+        if(images.length) {
+          S3.delete(images.map(img => img.image_url))
+          await Image.deleteMany({ "_id": { $in: gallery.images }})
+        }
+        await Gallery.deleteOne({ _id: gallery._id });
+        await deleteGallery(subdomain._id, gallery._id)
+        res.json(response.deleteGallery(`Successfuly deleted ${gallery.name} ` ))
+      } else {
+        throw { code: 404, message: "Gallery doesn't exist" }
+      }
     } catch (e) {
       res.json(response.buildError(e))
     }
